fix: 테스트용 계정 생성 api 버그 수정
