+++ user-service/src/main/java/com/t1/popcon/user/config/UserSecurityConfig.java
@@ -43,7 +43,8 @@ public class UserSecurityConfig extends CommonSecurityConfig {
                 ).permitAll()
                 .requestMatchers(
                         "/internal/users/**",
-                        "/internal/billing/**"
+                        "/internal/billing/**",
+                        "/users/internal/test-accounts/**"
                 ).permitAll()
                 .anyRequest().authenticated()
         );
