+++ user-service/src/main/java/com/t1/popcon/user/filter/InternalApiAuthFilter.java
@@ -18,6 +18,7 @@ import java.io.IOException;
 public class InternalApiAuthFilter extends OncePerRequestFilter {
 
     private static final String INTERNAL_API_PREFIX = "/internal/";
+    private static final String TEST_ACCOUNT_PREFIX = "/users/internal/test-accounts/";
     private static final String INTERNAL_SECRET_HEADER = "X-Internal-Secret";
 
     @Value("${internal.api-secret}")
@@ -25,7 +26,8 @@ public class InternalApiAuthFilter extends OncePerRequestFilter {
 
     @Override
     protected boolean shouldNotFilter(HttpServletRequest request) {
-        return !request.getRequestURI().startsWith(INTERNAL_API_PREFIX);
+        String uri = request.getRequestURI();
+        return !uri.startsWith(INTERNAL_API_PREFIX) && !uri.startsWith(TEST_ACCOUNT_PREFIX);
     }
 
     @Override
